backend/requirements.txt
fastapi
pandas
uvicorn
google-cloud-bigquery
google-cloud-aiplatform
google-generativeai
google-cloud-firestore
firebase-admin
pydantic
bcrypt
email-validator
redis
google-cloud-logging